Games/r2modman_utils/old_mods/pyproject.toml
[project]
name = "r2modman-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "cachier>=3.1.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
